Improve ease of use and fix typo.
It's nice to give the Composer user the exact Packagist spelling.  Sometimes the package name does not exactly match the Github project name.  This project is very new, so I wasn't even sure it was on Packagist.  Having to go search for it, even if easy to do, is a step that's not needed.  Nicer to just tell the user what they need to know here.